+++ lua/deja-view/init.lua
@@ -37,22 +37,6 @@ function M.restore()
     return
   end
 
-  -- Vim automatically remembers your last cursor position (but not your
-  -- view, which is why the plugin exists). The line/column memory is more
-  -- accurate than we can achieve through autocommands. Use it as a sanity
-  -- check.
-  --
-  -- Running `:edit` on the same buffer reloads it, but since there's no
-  -- event for deja-view to detect and save the view, we lose information
-  -- and incorrectly load the old one instead. Since the cursor position
-  -- doesn't change by `:edit`, we can simply ignore the event.
-  local last_line, last_col = unpack(vim.api.nvim_buf_get_mark(0, '"'))
-  if last_line > 1 and last_col > 0 then
-    if last_line ~= view.lnum and last_col ~= view.col then
-      return
-    end
-  end
-
   vim.fn.winrestview(view)
 end